# Speed up fpvec tests