This makes some changes to `Prio3Aes128FixedPointBoundedL2VecSum`-related tests to speed them up. First, the regression test from #409 is changed to run on a 5-element vector, rather than 1000. I confirmed that this test still fails if the fix to the FLP is reverted. (I also renamed various items throughout the test to reflect the changes.) This took 3.2s on my machine before, now it takes 0.03s.

Secondly, I split up `flp::types::fixedpoint_l2::tests::test_bounded_fpvec_sum_parallel` into four tests, so that they can be run in parallel. Previously, this took 0.86s on my machine, now the longest takes 0.61s. This is still one of the long poles, but I don't immediately see any other easy wins to take here.